refactor mock DAO file path
